Add sanity check for 'find' command. Git (the version control system) installer provides an option to replace Windows command-line tools with Unix ones (MSYS), which includes 'find' and can break our script. As the user can have their 'find' replaced unknowingly, we add the check and exits when 'find' behaves the Unix way.
This check depends on DOS/Windows 'find' erroring out when the first
argument is not quoted and not an option argument. Unix 'find' accepts it
as long as it's a valid path. So "find ." can tell the difference. To
suppress outputs and avoid useless recursion, we use syntax
"find . -prune >nul 2>nul".