Use "eol=\" on "FOR /F" in "DIR" statements.
The default comment ("eol") character in "FOR /F" is semicolon (;) rather than
none at all! This behaviour is not officially documented but is demonstrated
in various examples throughout the Internet. Unfortunately ";" is a valid
character in filenames and registry key names, this would makes filenames like
";foo.ext" not processed. Fix this by changing "eol" character to "\" which is
forbidden.

Note: Registry value names allow "\" characters.