Escape quotation mark in quote stripping code.
In theory this should prevent a syntax highlighter from complaining about
unpaired quotes. (But in practice most Batch syntax highlighters aren't smart
enough to recognize the escaped form ^" and handle that specially.) Escaping
characters is harmless for a !var:pattern=string! expansion due to its
'delayed' nature. So it's safe to apply this, and hope the future parsers and
highlighters will understand.