CodingStyle: Subroutines may now end with "EXIT /B" or non-EOF "GOTO"
Reason: Forcing the last line of subroutine be a GOTO :EOF could be too
restrictive for functions that return a value (exit code), or is forced
a loop that cannot be implemented with a FOR syntax (:continue_prompt
already serves as an example of this). If the subroutine must EXIT with
a code, having a "GOTO :EOF" after an "EXIT /B" is redundant. It's time
to loosen the rule.

Note that the last line of subroutine (which must be a "GOTO" or an
"EXIT /B") are NOT indented, in order to easily catch on unclosed
subroutines, which could be disasterous for Batch language and the
command interpreter.