CodingStyle: Device files now have ":" suffix.
Although the colon is optional now since DOS, it was a requirement to
identify the device in CP/M. CP/M filesystem does not forbid creation
of filenames like "NUL" or "CON" (no colon). And it's a good convention
even in DOS to always identify devices with ":" suffix anyway.

The '--move-subdir' option now also accepts "NUL:" value.

(Some drivers/protocols that allow accessing Unix filesystems in
Windows are not clever enough to filter or convert problematic
filenames like "nul" and will present the name as is to upper level
applications. Always requiring a ":" in our script can avoid some of
its problems, but not all.)