Replace "DANGER" word with "NOTICE" in messages.
Phase out the use of "DANGER" word in our program messages in favor of
well known keywords and established conventions.

Now we use only the keywords: "Notice", "Warning", "Error" and "Fatal
error".

The rule will be roughly:
"Notice": for expected conditions but significant enough for user to
take attention.
"Warning": for unexpected minor conditions which don't disrupt normal
operations/code flow.
"Error": for unexpected conditions that hinder the normal code flow,
causing the skip of some operations or making it work in the way that
could consider abnormal. (Sometimes the "ERROR:" keyword is not added
to the message if the message is intended to follow some error message
printed by another program.)
"Fatal error": for abnormalities that halt the script.