Make is_file_to_keep "short name" check work.
The 's' option in "%~sI" expansion is technically unspecified behavior
when used alone. Different cmd.exe versions make different behaviors on
that. (Win NT 4: %~sI expands to empty string; Win 2000 and later:
expands as if %~fsI is specified. Neither is what we want.)

Be specific. Specify "%~nxsI", which is no ambiguity.