Remove unneeded usebackq in "FOR /F" constructs.
This is a coding style change.
Reason: We will be introducing variable "FOR /F" options for these
"DIR /B" commands parsing, and the "usebackq" keyword can bring us
trouble when we reuse "FOR /F" options for other parts of code.
Let's limit usebackq usage to places where it's really necessary.