No more "No Disk" error dialogs when checking floppy drives.
Workaround the ugly pop-up behavior in Windows 2000 and XP.

This thing was the infamous "Abort, Retry, Ignore, Fail?" in DOS. In
Win9x it would present the same "Abort, Retry, Fail?" prompt in
terminal window. It's in WinNT that made it worse by popping up a
graphical dialog (even in Command Prompt environment) instead.

The pop-up really breaks consistency of how cmd.exe behaves on
accessing drives that are "not ready". Specifically it doesn't pop up a
dialog or block on CD-ROM drives but blocks on floppy drives. Bad, Bad
design.

Workaround this by using DIR command, which is the only option that
doesn't block the script (since NT 4).